Able to view the individual pages with the label but not the amount

+++ Fushionfastfood/Class/Food.swift
@@ -11,10 +11,10 @@ import Foundation
 class Food{
     
     var namE:String
-    var amounT:int_fast32_t
+    var amounT:Int16
     var imageName:String
      
-    init(name:String, amount:int_fast32_t, imagename:String){
+    init(name:String, amount:Int16, imagename:String){
          namE = name
         amounT = amount
         imageName = imagename
